rspirv2::binary

Trait WordWriter

pub trait WordWriter: Sized {
    // Required methods
    fn write(&mut self, word: Word);
    fn write_iter(&mut self, iter: impl IntoIterator<Item = Word>);

    // Provided methods
    fn write_op(&mut self, op: u16, len: usize) -> Result<(), EncodeError> { ... }
    fn inst_reserve(&mut self, len: usize) { ... }
}

A WordWriter allows you to Self::write Words into it and is usually backed by a Vec.

Implementations include Vec<Word> and WordCounter.

Required Methods

fn write(&mut self, word: Word)

Write a single Word

fn write_iter(&mut self, iter: impl IntoIterator<Item = Word>)

Write an Iterator of Words

Provided Methods

fn write_op(&mut self, op: u16, len: usize) -> Result<(), EncodeError>

Encode and write the opcode and len of an instruction.

Should not be overwritten.

fn inst_reserve(&mut self, len: usize)

Expect this many words to be written by the next instruction.

Usually called right before an instruction is emitted, with the length of the instruction, so expect this to be called quite often. If you’re writing into a Vec, prefer Vec::reserve over Vec::reserve_exact.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.