rspirv2::core

Module preamble

Re-exports

pub use super::extensions::*;

pub use super::grammar::*;

pub use super::inst::*;

pub use super::inst_class::*;

pub use super::inst_meta::*;

pub use super::inst_set::*;

pub use super::operand_kinds::*;

pub use super::operands::*;

pub use crate::custom_inst::*;

pub use crate::dis::*;

pub use crate::*;

Macros

bitflags

Generate a flags type.

Structs

CoreGrammar

DecodeError

Enumerant

A description of possible values for Category::BitEnum and Category::ValueEnum

ExtInstSetGrammar

Extension

A SPIR-V Extension

FixedLenComposer

Compose the OperandEncoding::FIXED_LEN from multiple maybe fixed len Operands

FnWriter

FormattedValue

Formatter

Configuration for formatting.

Grammar

A SPIR-V Grammar of any kind. There are only minor differences between the core SPIR-V specification and an extended instruction set, such as versioning.

IdMemorySemantics

A IdMemorySemantics is a reference to a IdResult of another operation.

IdRef

A IdRef is a reference to a IdResult of another operation.

IdResult

A SPIR-V “Result ID”.

IdResultAlloc

A sharable atomic IdResultAlloc, just Clone to share it

IdResultType

A IdResultType is a reference to a IdResult of another operation.

IdResultWriter

IdScope

A IdScope is a reference to a IdResult of another operation.

InstClass

Informational metadata about the class of instruction

InstDis

InstMeta

InstOffset

An offset to some instruction, a wrapper around usize.

InstReader

Reader for a single instruction

InstRef

A reference to a valid instruction encoded in a slice of Words.

LiteralConst

A LiteralContextDependentNumber, or LiteralConst for short since it’s only used by OpConstant (and OpSpecConstant) instructions.

LiteralExtInstInteger

The Instruction ID from an extended instruction set, backed by a 32bit integer.

LiteralFloat

A float literal as defined by SPIR-V spec: simply an f32. It’s represented as a u32 to ensure the bit pattern isn’t changing.

LiteralInteger

An integer literal as defined by SPIR-V spec: a 32bit integer.

LiteralSpecConstantOpInteger

The Instruction ID for an OpSpecConstantOp, backed by a 32bit integer.

LiteralString

A SPIR-V String literal. Defined as a sequence of UTF-8, so we can just use an ordinary String.

OperandDis

OperandDisContext

Context object for disassembly generation of operands within an instruction

OperandKind

Specifies possible Operand values, see Category variants.

OperandReader

Read operand Words from an instruction

OperandSpecMeta

An OperandKind of an instruction with a repetition Quantifier

ParameterizedBitmask

SmallVec

A Vec-like container that can store a small number of elements inline.

WordCounter

Counts the amount of Words that were written, discarding the words itself

WordSliceWriter

Enums

Category

The category of an OperandKind

ConstFmt

How the untyped constant value of a LiteralConst should be formatted

DecodeErrorKind

EncodeError

LiteralStringEscape

Describes how to escape string sequences

Quantifier

How many times to repeat something?

Constants

ID_RESULT_COLOR

ID_RESULT_MAX

A reasonable upper bounds that ensures subsequent allocations by potentially multiple threads all fail

OPERAND_KIND_ID_MEMORY_SEMANTICS

OPERAND_KIND_ID_REF

OPERAND_KIND_ID_RESULT

OPERAND_KIND_ID_RESULT_TYPE

OPERAND_KIND_ID_SCOPE

OPERAND_KIND_LITERAL_CONTEXT_DEPENDENT_NUMBER

OPERAND_KIND_LITERAL_EXT_INST_INTEGER

OPERAND_KIND_LITERAL_FLOAT

OPERAND_KIND_LITERAL_INTEGER

OPERAND_KIND_LITERAL_SPEC_CONSTANT_OP_INTEGER

OPERAND_KIND_LITERAL_STRING

PARAMETERIZED_BITMASK_REQUIRES_FIXED_LEN

Traits

AnyCapability

Inst

InstEncoding

MaybeIdResult

A type that may be an IdResult or ().

Operand

A SPIR-V operand. The associated const Self::KIND links to it’s OperandKind.

OperandEncoding

Something that can be decoded from or encoded to SPIR-V, not necessarily a full Operand.

ParameterizedBitmaskBits

WordWriter

A WordWriter allows you to Self::write Words into it and is usually backed by a Vec.

Type Aliases

LiteralContextDependentNumber

ZeroOrMore

ZeroOrOne