rspirv2::core::preamble

Struct InstRef

pub struct InstRef<'a, ISA>
where
    ISA: InstEncoding,
{ /* private fields */ }

A reference to a valid instruction encoded in a slice of Words.

Call Self::get to get the underlying instruction, as it unfortunately can’t implement Deref.

Safety

The referenced words must be an encoded instruction that is valid within the ISA instruction set. Encountering an invalid instruction may panic.

Implementations

impl<'a, ISA> InstRef<'a, ISA>
where ISA: InstEncoding,

pub fn from_words_unchecked( words: &'a [Word], ) -> Result<InstRef<'a, ISA>, DecodeError>

pub fn get(&self) -> ISA