rspirv2::core::preamble

Trait OperandEncoding

pub unsafe trait OperandEncoding: Sized + Debug {
    const FIXED_LEN: Option<usize>;

    // Required methods
    fn encode(&self, writer: &mut impl WordWriter) -> Result<(), EncodeError>;
    fn decode(reader: &mut OperandReader<'_>) -> Result<Self, DecodeError>;
    fn dis_fmt(
        &self,
        f: &mut Formatter<'_>,
        ctx: &OperandDisContext<'_>,
    ) -> Result<(), Error>;

    // Provided methods
    fn word_len(&self) -> usize { ... }
    fn decode_last(reader: &mut OperandReader<'_>) -> Result<Self, DecodeError> { ... }
    fn dis<'a>(&'a self, ctx: &'a OperandDisContext<'_>) -> OperandDis<'a, Self> { ... }
}

Something that can be decoded from or encoded to SPIR-V, not necessarily a full Operand.

Both Option and Vec implement OperandEncoding but not Operand. This allows for an easier representation of OperandSpecMetas with Quantifier of Quantifier::ZeroOrOne (Option) and Quantifier::ZeroOrMore (Vec).

Safety

Self::encode must WordWriter::write exactly Self::word_len many words.

Self::decode must OperandReader::pull (or Iterator::next) exactly Self::word_len many words.

If Self::FIXED_LEN is Some, it must equal the computed Self::word_len.

These constraints should only be validated with cfg!(debug_assertions), which are enabled by default in debug builds.

Required Associated Constants

const FIXED_LEN: Option<usize>

The fixed length of the Operand, or None if it’s variable length. Specifying this is an optimization for operand length calculation. See the safety contract in OperandEncoding.

Required Methods

fn encode(&self, writer: &mut impl WordWriter) -> Result<(), EncodeError>

Encode this Operand to a sequence of words.

fn decode(reader: &mut OperandReader<'_>) -> Result<Self, DecodeError>

Parse the Operand from the supplied Iterator of words, advancing it in the process.

fn dis_fmt( &self, f: &mut Formatter<'_>, ctx: &OperandDisContext<'_>, ) -> Result<(), Error>

Disassemble this operand to the supplied Formatter. Prefer Self::dis over calling this.

Provided Methods

fn word_len(&self) -> usize

The length of the operand in words.

By default, returns Self::FIXED_LEN if it is Some, or computes it by Self::encodeing with WordCounter. If encoding your type could be expensive, we recommend overwriting this implementation, although LLVM should be able to optimize most encoders into a simple expression.

Several debug assertions to check whether the length is correct are masked behind cfg!(debug_assertions).

fn decode_last(reader: &mut OperandReader<'_>) -> Result<Self, DecodeError>

Like Self::decode, but the Operand is guaranteed to be the last one of this Instruction. This is used by LiteralInteger to consume all the remaining words and not have to calculate the exact size of a type.

See LiteralInteger for details.

fn dis<'a>(&'a self, ctx: &'a OperandDisContext<'_>) -> OperandDis<'a, Self>

Disassemble this operand

Returns a type that impl Display, so it’s usable with format!:

let my_op = IdResult(Word(42));
let dis = format!("OpMyInst {}", my_op.dis());
assert_eq!(dis, "OpMyInst %42");

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.