rspirv2

Module custom_inst

custom instructions

Structs

OpSwitch

OpSwitch requires a custom implementation, as it is the only SPIR-V instruction which decoding is context-dependent and not resolvable by grabbing the remaining words available.

OpSwitchResolvedTarget

An OpSwitch where the word length of the target has been resolved.

Enums

OpSwitchError

OpSwitchTarget

Target of an OpSwitch

OpSwitchTargetLen

Length of the target literal of an OpSwitch in words