rspirv2::custom_inst

Struct OpSwitch

pub struct OpSwitch {
    pub selector: IdRef,
    pub default: IdRef,
    pub target: OpSwitchTarget,
}

OpSwitch requires a custom implementation, as it is the only SPIR-V instruction which decoding is context-dependent and not resolvable by grabbing the remaining words available.

The target of the OpSwitch are ZeroOrMore tuples of LiteralConst and IdRef. The selector is compared with every LiteralConst in target, and if one matches, jumps to the label referenced by the IdRef. Otherwise, it jumps to the default label.

The problem is that without context, it’s unknown whether LiteralConst should consume one or two words, for 32bit (or lower) or 64bit literals. Unlike crate::core::inst::OpConstant, you can’t just consume the remaining words and based on that assume the constant is one or two words (see LiteralConst docs). For example, with 6 words remaining, you couldn’t tell whether target has 3 32bit entries or 2 64bit entries.

So we need context to resolve the type of selector to correctly decode target. This library has been specifically designed so that instruction encoding and decoding is context-free, which greatly simplifies many aspects of it, so we don’t want to abandon those principles. Instead, we choose to special-case OpSwitch with our custom instruction set API and defer the decoding of target until it is “used”.

The target operand is replaced by OpSwitchTarget, an enum that can store target as Unresolved in plain words or as Resolved to a specific OpSwitchTargetLen of either One or Two word literals. While u64 requires a Two word literal, u32 (also u16 and u8) are stored in a One word literal. Use OpSwitch::resolve to specify the word size and convert an Unresolved target to a Resolved one, returning a &mut reference to the latter. This will also store the OpSwitchTargetLen used for decoding in the type, to validate that future calls to resolve must use the same word len. Note that encoding and decoding the OpSwitch instruction makes it lose that information.

Fields
selector: IdRefdefault: IdReftarget: OpSwitchTarget
Implementations

impl OpSwitch

pub fn resolve( &mut self, len: OpSwitchTargetLen, ) -> Result<&mut OpSwitchResolvedTarget, OpSwitchError>

Resolve this OpSwitch instruction to have a One word or Two words literal. Stores the decoded form and word len in the struct to make future retrieval fast and validates against mismatched len.

pub fn resolve_ref( &self, len: OpSwitchTargetLen, ) -> Result<Cow<'_, OpSwitchResolvedTarget>, OpSwitchError>

Resolve this OpSwitch instruction to have a One word or Two words literal. Does not store the decoded form in the type, making validation on accidental reinterpretation possible.