rspirv2::dis

Trait InstSetDisCtx

pub trait InstSetDisCtx: InstEncoding {
    // Required method
    fn add_context(slice: &RawInstSlice, ctx: &mut DisContext);
}

An instruction set that provides additional context information for disassembly.

Required Methods

fn add_context(slice: &RawInstSlice, ctx: &mut DisContext)

Add context to the supplied DisContext by modifying the various public members of it.

Only supplies a RawInstSlice instead of a full InstSlice that has been error checked, as to allow disassembly of partially corrupt instructions. Any DecodeError that arrises during decode should be silently ignored. We recommend using slice.iter().try_decode::<CoreInstSet>().skip_errors() to decode instructions from the RawInstSlice.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.