rspirv2::inst

Trait InstEncoding

pub trait InstEncoding:
    Sized
    + Debug
    + Eq {
    type IdResult: MaybeIdResult;
    type IdResultType: MaybeIdResult;

    // Required methods
    fn id_result(&self) -> Self::IdResult;
    fn id_result_type(&self) -> Self::IdResultType;
    fn encode(&self, writer: &mut impl WordWriter) -> Result<(), EncodeError>;
    fn decode(reader: InstReader<'_>) -> Result<Self, DecodeError>;
    fn dis_fmt(
        &self,
        f: &mut Formatter<'_>,
        ctx: &DisContext,
    ) -> Result<(), Error>;

    // Provided methods
    fn name() -> &'static str { ... }
    fn try_decode(reader: InstReader<'_>) -> Result<Option<Self>, DecodeError> { ... }
    fn dis<'a>(&'a self, ctx: &'a DisContext) -> InstDis<'a, Self> { ... }
}

Required Associated Types

type IdResult: MaybeIdResult

IdResult is either an IdResult or (), depending on whether this Instruction has an IdResult.

type IdResultType: MaybeIdResult

IdResult is either an IdResult or (), depending on whether this Instruction has an crate::operand::IdResultType.

Required Methods

fn id_result(&self) -> Self::IdResult

Query the potential IdResult of this Instruction, or () if it has none.

fn id_result_type(&self) -> Self::IdResultType

Query the potential IdResult of this Instruction, or () if it has none.

fn encode(&self, writer: &mut impl WordWriter) -> Result<(), EncodeError>

Encode this instruction to a WordWriter

fn decode(reader: InstReader<'_>) -> Result<Self, DecodeError>

Decode this instruction from an InstReader, error when opcode is unknown.

See Self::try_decode for a variant that returns None when opcode is unknown.

fn dis_fmt(&self, f: &mut Formatter<'_>, ctx: &DisContext) -> Result<(), Error>

Disassemble this instruction to the supplied Formatter f

Provided Methods

fn name() -> &'static str

Name of the instruction set, for debug printing

fn try_decode(reader: InstReader<'_>) -> Result<Option<Self>, DecodeError>

Try to decode this instruction from an InstReader, return None when the opcode is unknown.

See Self::decode for a variant that errors when opcode is unknown.

fn dis<'a>(&'a self, ctx: &'a DisContext) -> InstDis<'a, Self>

Disassemble this instruction, returns a type that impl Display you may use in format!.

You shouldn’t overwrite this function but Self::dis_fmt instead.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.