rspirv2

Module meta

Rust structs containing metadata information about SPIR-V instructions, directly deserialized from the SPIR-V grammar JSON.

These structs are a derivative of:

rspirv

spirt

Structs

CoreGrammar

Enumerant

A description of possible values for Category::BitEnum and Category::ValueEnum

ExtInstSetGrammar

Extension

A SPIR-V Extension

Grammar

A SPIR-V Grammar of any kind. There are only minor differences between the core SPIR-V specification and an extended instruction set, such as versioning.

InstClass

Informational metadata about the class of instruction

InstMeta

OperandKind

Specifies possible Operand values, see Category variants.

OperandSpecMeta

An OperandKind of an instruction with a repetition Quantifier

Enums

Category

The category of an OperandKind

Quantifier

How many times to repeat something?

Traits

AnyCapability