rspirv2::module

Struct Module

pub struct Module<ISA>
where
    ISA: InstEncoding,
{
    pub header: Option<SpirvHeader>,
    pub inst: InstVec<ISA>,
}

A SPIR-V Module is an InstVec with an optional SpirvHeader.

See https://registry.khronos.org/SPIR-V/specs/unified1/SPIRV.html#_physical_layout_of_a_spir_v_module_and_instruction

Fields
header: Option<SpirvHeader>inst: InstVec<ISA>
Implementations

impl<ISA> Module<ISA>
where ISA: InstEncoding,

pub fn from_bytes(bytes: &[u8]) -> Result<Module<ISA>, ParseError>

Parse a SPIR-V module from bytes, endianness is automatically detected and instructions checked for validity.

pub fn from_bytes_unchecked(bytes: &[u8]) -> Result<Module<ISA>, ParseError>

Parse a SPIR-V module from bytes, endianness is automatically detected.

Instructions are not checked for validity, and you may get panics due to invalid instructions later on.

pub fn from_words_maybe_header( words: &[Word], ) -> Result<Module<ISA>, ParseError>

Parse a SPIR-V Module from a word slice that may or may not contain a SpirvHeader.

pub fn from_words_maybe_header_unchecked( words: &[Word], ) -> Result<Module<ISA>, ParseError>

Parse a SPIR-V Module from a word slice that may or may not contain a SpirvHeader.

Instructions are not checked for validity, and you may get panics due to invalid instructions later on.

pub fn write_words(&self, writer: &mut impl WordWriter)

Write this module including the header to a WordWriter

pub fn write_bytes(&self, writer: &mut impl Write) -> Result<(), Error>

Write this module including the header to a std::io::Write, which could be a std::fs::File