rspirv2::operand

Trait ParameterizedBitmaskBits

pub trait ParameterizedBitmaskBits:
    Copy
    + OperandEncoding
    + Flags<Bits = u32> {
    const BIT_TO_EXTRA_LEN: &'static [usize];

    // Provided methods
    fn bits_cnt() -> u32 { ... }
    fn assert_no_extra(bit: u32) { ... }
    fn extra_offset(&self, bit: u32) -> usize { ... }
    fn extra_offset_len(&self, bit: u32) -> (usize, usize) { ... }
    fn extra_total_len(&self) -> usize { ... }
}

Required Associated Constants

const BIT_TO_EXTRA_LEN: &'static [usize]

Provided Methods

fn bits_cnt() -> u32

How many bits this bitmask has

fn assert_no_extra(bit: u32)

fn extra_offset(&self, bit: u32) -> usize

Offset of the extra words for the given bit

fn extra_offset_len(&self, bit: u32) -> (usize, usize)

Offset and len of the extra words for the given bit

fn extra_total_len(&self) -> usize

total len of words

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.