rspirv2::slice

Struct InstSlice

pub struct InstSlice<ISA>
where
    ISA: InstEncoding,
{ /* private fields */ }

An InstSlice works like a &[ISA], but stores instructions in the variable-sized SPIR-V binary form to safe on memory.

Instruction sets are a giant enum of all instructions, their size is determined by largest instruction plus a discriminant. The largest instructions in the core ISA are made from 10-15 operands, the enum of all instructions is currently 80 bytes large. If you’d put that directly into a Vec, you’d be paying 80 bytes per instruction, even though the most common instructions only need 2-4 operands or 12-20 bytes, wasting a lot of memory.

This InstVec stores instructions in their binary SPIR-V form, which is variable sized, allowing us to save a lot of memory and improving cache locality. To make it as convenient to use as a Vec<ISA>, instructions are automatically encoded and decoded on the fly. To remain performant, decoding of instructions should be as cheap as possible. Borrowing from the Vec directly hasn’t been implemented yet, so some operands like LiteralString allocate a string and copy out the bytes. LiteralConst uses a SmallVec, so the allocation shouldn’t happen in pretty much all cases.

However, storing instructions in a variable-sized way brings the same disadvantages as UTF-8 characters in rust strings. You can’t arbitrarily index into an InstSlice like you can with a regular slice, as only indices pointing to the beginning of an instruction are valid. You also can’t easily replace instructions in the middle of the stream, as changing the size of an instruction requires you to move all following instructions. Although, you can remove instructions by filling them with OpNop, but this hasn’t been implemented.

Safety

The inner slice of words is assumed to contain valid instructions of the generic ISA Instruction Set. May panic if instructions fail to decode, but will not lead to UB, allowing Self::from_words_unchecked to be safe.

Implementations

impl<ISA> InstSlice<ISA>
where ISA: InstEncoding,

pub fn from_words(words: &[Word]) -> Result<&InstSlice<ISA>, DecodeError>

Create a new InstSlice from a &[Word] safely, by verifying the instructions to be valid in the ISA.

pub const fn from_words_unchecked(words: &[Word]) -> &InstSlice<ISA>

Create a new InstSlice from a &[Word] without checking whether the words contain valid instructions.

See crate::vec::InstVec # Safety

pub fn from_raw(raw: &RawInstSlice) -> Result<&InstSlice<ISA>, DecodeError>

Create a new InstSlice from a RawInstSlice safely, by verifying the instructions to be valid in the ISA.

pub const fn from_raw_unchecked(raw: &RawInstSlice) -> &InstSlice<ISA>

Create a new InstSlice from a RawInstSlice without checking whether the words contain valid instructions.

Safety

See InstSlice #Safety

pub const fn empty() -> &'static InstSlice<ISA>

An InstSlice with 0 instructions

pub const fn as_raw(&self) -> &RawInstSlice

View self as a RawInstSlice

pub const fn iter_ref(&self) -> InstRefIter<'_, ISA>

Iterate over instruction references (InstRef) without decoding the instruction itself

pub const fn iter(&self) -> InstIter<'_, ISA>

Iterate over all instructions

impl<ISA> InstSlice<ISA>
where ISA: InstEncoding,

pub fn get_ref(&self, offset: InstOffset) -> Option<InstRef<'_, ISA>>

Get instruction at offset as an InstRef, return None when index is invalid

pub fn get(&self, offset: InstOffset) -> Option<ISA>

Get instruction at offset, return None when index is invalid

pub fn index_ref(&self, offset: InstOffset) -> InstRef<'_, ISA>

Get instruction at offset as an InstRef, panic when index is invalid

pub fn index(&self, offset: InstOffset) -> ISA

Get instruction at offset, panic when index is invalid

pub fn slice<R>(&self, index: R) -> Option<&InstSlice<ISA>>
where R: RangeBounds<InstOffset>,

Slice this InstSlice

impl<ISA> InstSlice<ISA>
where ISA: InstSetDisCtx,

pub fn dis(&self, ctx: impl IntoDisContext) -> DisInstSlice<'_, ISA>

disassemble