rspirv2::slice

Struct RawInstSlice

#[repr(transparent)]
pub struct RawInstSlice(pub [Word]);

A slice of SPIR-V instructions of an arbitrary instruction set.

Compared to InstSlice, does not specify the instruction set nor require the inner words to be valid instructions in said instruction set.

Tuple Fields
0: [Word]
Implementations

impl RawInstSlice

pub const fn from_words(words: &[Word]) -> &RawInstSlice

pub const fn empty() -> &'static RawInstSlice

An InstSlice with 0 instructions

pub const fn as_words(&self) -> &[Word]

Returns the underlying slice of words

pub const fn iter(&self) -> RawInstRefIter<'_>

Iterate over instruction references (InstRef) without decoding the instruction itself

pub fn verify_valid_in_isa<ISA>(&self) -> Result<(), DecodeError>
where ISA: InstEncoding,

Verify whether the instructions are valid within the generic ISA instruction set

impl RawInstSlice

pub fn dis<ISA>(&self, ctx: impl IntoDisContext) -> DisInstSlice<'_, ISA>
where ISA: InstSetDisCtx,

disassemble