rspirv2::vec

Struct InstVec

pub struct InstVec<ISA>
where
    ISA: InstEncoding,
{ /* private fields */ }

An InstVec works like a Vec<ISA>, but stores instructions in the variable-sized SPIR-V binary form to safe on memory size.

See InstSlice for details.

Implementations

impl<ISA> InstVec<ISA>
where ISA: InstEncoding,

pub const fn new() -> InstVec<ISA>

Create a new empty InstVec

pub fn from_words(words: Vec<Word>) -> Result<InstVec<ISA>, DecodeError>

Create a new InstVec from a Vec safely, by verifying the instructions within to be valid for the ISA

pub fn from_raw(raw: RawInstVec) -> Result<InstVec<ISA>, DecodeError>

Create a new InstVec from a RawInstVec safely, by verifying the instructions within to be valid for the ISA

pub fn from_words_unchecked(words: Vec<Word>) -> InstVec<ISA>

Create a new InstVec from a Vec of words without checking whether the words contain valid instructions.

See InstVec # Safety

pub fn from_raw_unchecked(raw: RawInstVec) -> InstVec<ISA>

Create a new InstVec from a RawInstVec without checking whether the words contain valid instructions.

See InstVec # Safety

pub fn push(&mut self, inst: impl Into<ISA>) -> InstOffset

Appends an instruction to the back of the InstVec. See Vec::push.

pub fn push_inst<I>(&mut self, inst: I) -> <I as InstEncoding>::IdResult
where I: Inst + Into<ISA>,

Appends an instruction to the back of the InstVec. See Vec::push.

pub fn append(&mut self, other: &mut InstVec<ISA>)

Moves all the instruction of other into self, leaving other empty. See Vec::append.

pub fn into_raw(self) -> RawInstVec

Turn into the inner Vec of words

pub fn into_vec(self) -> Vec<Word>

Turn into the inner Vec of words

pub const fn as_slice(&self) -> &InstSlice<ISA>

View this InstVec as a InstSlice

pub const fn as_raw_slice(&self) -> &RawInstSlice

View this InstVec as a RawInstSlice