rspirv2::vec

Struct RawInstVec

pub struct RawInstVec(pub Vec<Word>);

A slice of SPIR-V instructions of an arbitrary instruction set.

Compared to InstVec, does not specify the instruction set nor require the inner words to be valid instructions in said instruction set.

Tuple Fields
0: Vec<Word>
Implementations

impl RawInstVec

pub const fn new() -> RawInstVec

Create a new empty RawInstVec

pub fn from_words(words: Vec<Word>) -> RawInstVec

Create a new RawInstVec from a Vec

pub fn push(&mut self, inst: impl InstEncoding) -> InstOffset

Appends an instruction to the back of the RawInstVec, like Vec::push.

pub fn push_inst<I>(&mut self, inst: I) -> <I as InstEncoding>::IdResult
where I: Inst,

Appends an instruction to the back of the RawInstVec. See Vec::push.

pub fn append(&mut self, other: &mut RawInstVec)

Moves all the instruction of other into self, leaving other empty. See Vec::append.

pub fn into_vec(self) -> Vec<Word>

Turn into the inner Vec of words

pub const fn as_slice(&self) -> &RawInstSlice

View this RawInstVec as a RawInstSlice